Constrain scanpid.sh and busywait.sh processes to CPU 2
And place the script and mapper process on CPU 0.

+++ proc-vs-map.sh
@@ -2,22 +2,29 @@
 
 duration=${1-5}
 ncpus="`lscpu | grep '^CPU(s):' | awk '{ print $2 }'`"
+if test "$ncpus" -lt 2
+then
+	echo Need at least two CPUs, and ncpus = $ncpus
+	exit 1
+fi
+nbusytasks=5
 
 T=/tmp/proc-vs-map.sh.$$
 trap 'rm -rf $T' 0 2
 mkdir $T
 
-echo Starting test `date`
+echo Starting ${duration}-second test at `date`.
 
-./mapper --duration $duration > $T/mapper.out &
+taskset -p 0x1 $$
+taskset -c 0 ./mapper --duration $duration > $T/mapper.out &
 mapper_pid=$!
-echo "(Expect one complaint per CPU from scanpid.sh about /proc/$mapper_pid/smaps at end of test.)"
+echo "(Expect $nbusytasks complaints from scanpid.sh about /proc/$mapper_pid/smaps at end of test.)"
 
 busy_pids=
 i=0
-while test $i -lt $ncpus
+while test $i -lt $nbusytasks
 do
-	taskset -c $i nice -n 15 ./scanpid.sh $mapper_pid &
+	taskset -c 2 nice -n 15 ./scanpid.sh $mapper_pid &
 	busy_pids="$busy_pids $!"
 	i=$((i+1))
 done
@@ -25,9 +32,9 @@ done
 sleep 1
 
 i=0
-while test $i -lt $ncpus
+while test $i -lt $nbusytasks
 do
-	taskset -c $i nice -n 10 ./busywait.sh &
+	taskset -c 2 nice -n 10 ./busywait.sh &
 	busy_pids="$busy_pids $!"
 	i=$((i+1))
 done
